Add persistent Resonance progression and skill data

+++ src/main/java/kmoth/resonance/skill/PulseStepData.java
@@ -4,6 +4,6 @@ public class PulseStepData {
 
     public int cooldown_seconds;
     public double speed_multiplier;
-    public int duration_seconds;
+    public double duration_seconds;
     public int skill_point_cost;
 }
